requirements.txt
pandas
numpy
matplotlib
scikit-learn
xgboost
fastapi
uvicorn[standard]
streamlit
sqlalchemy
psycopg2-binary
confluent-kafka
websockets
python-dotenv
pydantic
requests
